#17110 Move promise utils to base test common layer
